linketron/requirements.txt
aiogram
google-generativeai
python-dotenv
openai
requests
google-genai
Pillow
linkedin-api